requirements.txt
aiohttp==3.0.1
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==2.0.0
attrs==17.4.0
Babel==2.4.0
beautifulsoup4==4.6.0
blinker==1.4
boto3==1.17.76
botocore==1.20.76
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
debtcollector==1.13.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.5
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
docutils==0.14
funcsigs==1.0.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==4.6.1
iso8601==0.1.11
jmespath==0.10.0
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
keystoneauth1==3.4.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==1.0
monotonic==1.0
msgpack==0.5.6
multidict==4.1.0
netaddr==0.7.19
netifaces==0.10.4
numpy==1.17.2
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
oslo.config==5.2.0
oslo.i18n==3.19.0
oslo.serialization==2.24.0
oslo.utils==3.35.0
pbr==3.1.1
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
positional==1.1.1
protobuf==3.14.0
pycairo==1.16.2
pycodestyle==2.3.1
pycparser==2.18
pycrypto==2.6.1
pycups==1.9.73
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-keystoneclient==3.15.0
python-swiftclient==3.5.0
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
rfc3986==0.3.1
roman==2.0.0
s3transfer==0.4.2
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
sqlparse==0.4.1
ssh-import-id==5.7
stevedore==1.28.0
system-service==0.3
systemd-python==234
typing-extensions==3.10.0.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.3
virtualenv==16.7.5
wadllib==1.3.2
webencodings==0.5
wrapt==1.9.0
xkit==0.0.0
yarl==1.1.0
zipp==3.5.0
zope.interface==4.3.2
